+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## Version 110105.01
+
+- Added 11.1.5 support.
+
+### Unit Frames
+
+- Fixed an issue where the cast bar would break whenever you tried to empower a spell with 4+ stages.
+- Removed "[ls:sheepicon]" tag. It's way too inaccurate, it's time to lay the old sheep to rest.
+- Misc bug fixes and tweaks. 
+
 ## Version 110100.02
 
 ### Action Bars